Contract part three
